改 type

+++ ComicCatcherLib/ComicModels/ComicPage.cs
@@ -2,7 +2,7 @@
 
 public class ComicPage : ComicBaseProperty
 {
-    public int PageNumber { get; set; }
+    public long PageNumber { get; set; }
     public string PageFileName { get; set; }
     public string Refer { get; set; }
 }
